+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 jflow 0.4-pre-alpha
 ===================
  * Refactoring `instdata` and `trade` models. Not backward compatible
+ * Created django-piston API
+ * Added `jfsite` application site
 
 
 jflow 0.3   2010-03-15